+++ processor/src/test/resources/fixtures/org/mapstruct/ap/test/source/constants/ConstantsMapperImpl.java
@@ -33,12 +33,12 @@ public class ConstantsMapperImpl implements ConstantsMapper {
         constantsTarget.setIntValue( -03777777 );
         constantsTarget.setIntBoxed( 15 );
         constantsTarget.setLongValue( 0x7fffffffffffffffL );
-        constantsTarget.setLongBoxed( (long) 0xCAFEBABEL );
+        constantsTarget.setLongBoxed( 0xCAFEBABEL );
         constantsTarget.setFloatValue( 1.40e-45f );
         constantsTarget.setFloatBoxed( 3.4028235e38f );
         constantsTarget.setDoubleValue( 1e137 );
-        constantsTarget.setDoubleBoxed( (double) 0x0.001P-1062d );
-        constantsTarget.setDoubleBoxedZero( (double) 0.0 );
+        constantsTarget.setDoubleBoxed( 0x0.001P-1062d );
+        constantsTarget.setDoubleBoxedZero( 0.0 );
 
         return constantsTarget;
     }
